Move exists and loadData onto the player object
