+++ src/lib/auth/login.ts
@@ -19,7 +19,7 @@ export default class Login implements InputHandler {
     if (player.username === null) {
       data = data.trim().toLowerCase();
       player.username = data;
-      if (!this.playerExists(player)) {
+      if (!player.exists()) {
         player.sendData("Haven't seen you around here before.\n");
         player.sendData('What would you like for a password? ');
       }
@@ -28,45 +28,17 @@ export default class Login implements InputHandler {
       }
     }
     else {
-      const config = Config.getInstance();
-      const player_data = this.loadPlayer(player);
+      const player_data = player.loadData().playerData;
       console.debug('[debug] Loaded player data ', player_data);
       // Trim the trailing newline
       data = data.replace(/\r?\n|\r/g, "");
       if (checkPassword(data, player_data['password'])) {
         player.password = data;
-        player.playerData = player_data;
         player.sendData(`Welcome, ${player}!\n`);
       }
       else {
         player.sendData("Well that's just not right. Care to try again? ");
       }
     }
-  }
-
-  playerExists(player: Player): boolean {
-    const save_path = player.savePath;
-    if (fs.existsSync(save_path)) {
-      const stats = fs.lstatSync(save_path);
-      if (!stats.isFile()) {
-        console.error(`[error] Player save data exists, but is not a file?! (${save_path})`);
-        return false;
-      }
-      try {
-        fs.accessSync(save_path, fs.constants.W_OK);
-      }
-      catch (err) {
-        console.error(`[error] Player save file exists, but is not writable. (${save_path})`);
-        return false;
-      }
-      console.debug(`[debug] Found player save file: ${save_path}`);
-      return true;
-    }
-    console.debug(`[debug] Player save file does not exist: ${save_path}`);
-    return false;
-  }
-
-  loadPlayer(player: Player): Record<string, any> {
-    return toml.parse(fs.readFileSync(player.savePath).toString());
   }
 }
